% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{ask_gpt}
\alias{ask_gpt}
\title{ASK ChatGPT for some R help}
\usage{
ask_gpt(message, model = "text-davinci-003", maxtoken = 300, tempvalue = 0)
}
\arguments{
\item{message}{The prompt for GTP}

\item{model}{The GTP model}

\item{maxtoken}{Maximal numbers and tokens}

\item{tempvalue}{The temperature}
}
\value{
A character vector.
}
\description{
The function `ask_gpt` sends message to GTP; shows the results in the console, and copies
the result to your clipboard
}
