% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc.R
\docType{data}
\name{ccc}
\alias{ccc}
\title{Copy Cat Code:}
\format{
A data frame with 41 rows and 3 variables:
\describe{
  \item{package}{The R package name}
  \item{fct}{The function name}
  \item{code}{The code snippet}
  ...
}
}
\source{
\url{http://www.edgar-treischl.de/}
}
\usage{
ccc
}
\description{
A dataset containing minimal examples of code snippets
that run with implemented data.
}
\keyword{datasets}
