\name{copyCat-package}
\docType{package}
\title{
  What the package does (short line)
}
\description{
  Utilizes multi-core systems with the help of the foreach and doMC
  packages, degrades gracefully to single core systems.
}
\details{
  \tabular{ll}{
    Package: \tab readDepth\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2010-05-12\cr
    License: \tab Apache\cr
    LazyLoad: \tab yes\cr
  }
  

  ~~ An overview of how to use the package, including the most important ~~
  ~~ functions ~~
}
\author{
  Chris Miller
  
  Maintainer: Chris Miller <chrisamiller@gmail.com>
}
\references{
  TBA
}
\keyword{ package }
\examples{

}
