% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_code.R
\name{copycat}
\alias{copycat}
\title{copycat}
\usage{
copycat(x, data = CopyCatCode, run = FALSE)
}
\arguments{
\item{x}{A search string}

\item{data}{Data}

\item{run}{TRUE or FALSE}
}
\value{
A string
}
\description{
copycat
}
