% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_addin.R
\name{copycat_addin}
\alias{copycat_addin}
\title{Manage code snippets via the `copycat_addin()`.}
\usage{
copycat_addin(data = CopyCatCode)
}
\arguments{
\item{data}{A data frame}
}
\description{
The `copycat_addin()` starts a small app in the viewer
that shows the copycat or any data frame for Copycat. Just select an R package
and the corresponding code snippet. Then press
the button and the code will be insert into the current document at the
location of the cursor. The `copycat_addin()` is inspired by parsnip addin,
which write model specifications for you.
}
