% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_code.R
\name{copycat_code}
\alias{copycat_code}
\title{copycat_code}
\usage{
copycat_code(x, data = CopyCatCode)
}
\arguments{
\item{x}{A search string}

\item{data}{Data}
}
\value{
A string
}
\description{
copycat_code
}
