% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_help.R
\name{copycat_description}
\alias{copycat_description}
\title{Copy the function description}
\usage{
copycat_description(pkg, fn, html = FALSE)
}
\arguments{
\item{pkg}{A search string}

\item{fn}{A search string}

\item{html}{HTML or txt}
}
\value{
Function description
}
\description{
The function returns the function description.
}
