% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist.R
\name{copycat_gistfiles}
\alias{copycat_gistfiles}
\title{Explore GitHub Gist files}
\usage{
copycat_gistfiles()
}
\value{
Data frame with file names.
}
\description{
The function `copycat_gistfiles()` returns a list with GitHub Gists.
}
