% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist.R
\name{copycat_gists.create}
\alias{copycat_gists.create}
\title{Create Github Gist files}
\usage{
copycat_gists.create(name, code, description)
}
\arguments{
\item{name}{The name of the Gist file.}

\item{code}{String with the code of the Gist.}

\item{description}{Code description.}
}
\value{
Message if successful.
}
\description{
The function `copycat_gists.create()` creates GitHub Gists.
}
