% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist.R
\name{copycat_gists.del}
\alias{copycat_gists.del}
\title{Delete Github Gist files}
\usage{
copycat_gists.del(id)
}
\arguments{
\item{id}{The id of the Gist file (see copycat_gistfiles).}
}
\value{
Message if successful.
}
\description{
The function `copycat_gists.del()` deletes GitHub Gists.
}
