% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_copy.R
\name{copycat_git}
\alias{copycat_git}
\title{copycat_git}
\usage{
copycat_git(file)
}
\arguments{
\item{file}{A character vector.}
}
\value{
A character vector.
}
\description{
Copy a file of a github repository. It expects that
`git_setup` is set before running, a data frame with column for the
`author`, the `repository`, and the `branch` name; `copycat_git` copies
the code to your clipboard.
}
