% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_copy.R
\name{copycat_gitcode}
\alias{copycat_gitcode}
\title{Inspect R code from a Github repository.}
\usage{
copycat_gitcode(file)
}
\arguments{
\item{file}{A character vector.}
}
\value{
A character vector.
}
\description{
Inspect the code of a file from a github repository. It expects that
`git_setup` is set before running; a data frame with a column for the
`author`, the `repository`, and the `branch` name; `copycat_gitcode` returns
the code as character.
}
