% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_copy.R
\name{copycat_gitplot}
\alias{copycat_gitplot}
\title{copycat_gitplot}
\usage{
copycat_gitplot(file)
}
\arguments{
\item{file}{A character vector.}
}
\value{
A character vector.
}
\description{
Copy and run a file of a github repository. It expects that
`git_setup` is set before running; a data frame with a column for the
`author`, the `repository`, and the `branch` name; `copycat_gitplot` copies
the code and sends it your console.
}
