% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_copy.R
\name{copycat_gitsearch}
\alias{copycat_gitsearch}
\title{Search R files within a GitHub directory.}
\usage{
copycat_gitsearch(author, repository, branch = "main")
}
\arguments{
\item{author}{A character vector.}

\item{repository}{A character vector.}

\item{branch}{A character vector.}
}
\value{
A character vector.
}
\description{
The function `copycat_gitsearch` returns the list of R files in a
Github repository. It expects `author`, `repository`, and `branch` name to
print the list of R files on Github. If all is left out, it returns the
content of example repository.
}
