% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_help.R
\name{copycat_help}
\alias{copycat_help}
\title{copycat_help}
\usage{
copycat_help(pkg, fn)
}
\arguments{
\item{pkg}{A search string}

\item{fn}{A search string}
}
\value{
A string
}
\description{
copycat_help
}
