% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{copycat_helpcode}
\alias{copycat_helpcode}
\title{copycat_helpcode}
\usage{
copycat_helpcode(pkg, fn)
}
\arguments{
\item{pkg}{A search string}

\item{fn}{A search string}
}
\value{
A string
}
\description{
The copycat_helpcode function copies
the code from the help file (examples) from an R package
the line types in R
}
