% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_help.R
\name{copycat_helpscript}
\alias{copycat_helpscript}
\title{Create a new script with the code from the help file.}
\usage{
copycat_helpscript(pkg, fn, path = NULL)
}
\arguments{
\item{pkg}{A search string}

\item{fn}{A search string}

\item{path}{A search string}
}
\value{
A string
}
\description{
Create a new script with the code from the help file.
}
