% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_help.R
\name{copycat_inserthelp}
\alias{copycat_inserthelp}
\title{copycat_helpfile}
\usage{
copycat_inserthelp(pkg, fn, path = NULL)
}
\arguments{
\item{pkg}{A search string}

\item{fn}{A search string}

\item{path}{A search string}
}
\value{
A string
}
\description{
copycat_helpfile
}
