% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_code.R
\name{copycat_package}
\alias{copycat_package}
\title{Find the corresponding package of a function.}
\usage{
copycat_package(x, data = CopyCatCode)
}
\arguments{
\item{x}{A search string for the code snippet}

\item{data}{A data frame}
}
\value{
A string with the package name and the function sends
the code to load the library to the console.
}
\description{
Use `copycat_package()` expects a text string to search for
corresponding package name. It expects that the data frame has three columns.
A column with the package name (package), a function (fun), and one with
the code (code) to copy. If data = NULL, the package search in the
the internal data `CopyCatCode`.
}
