% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{copycat_random}
\alias{copycat_random}
\title{Run code from a random package and function}
\usage{
copycat_random()
}
\value{
A string
}
\description{
The function copies
an example code from a help file randomly and sends
it to the console
}
