% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{copycat_runhelp}
\alias{copycat_runhelp}
\title{copycat_runhelp}
\usage{
copycat_runhelp(pkg, fn)
}
\arguments{
\item{pkg}{A search string}

\item{fn}{A search string}
}
\value{
A string
}
\description{
copycat_runhelp
}
