% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_vignettes.R
\name{copycat_vignette}
\alias{copycat_vignette}
\title{Copy the code of a vignette.}
\usage{
copycat_vignette(pkg, vig)
}
\arguments{
\item{pkg}{A search string}

\item{vig}{A search string}
}
\value{
A string
}
\description{
Copy the code of a vignette.
}
