% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_vignettes.R
\name{copycat_vigscript}
\alias{copycat_vigscript}
\title{Create a new script with the code from the vignette.}
\usage{
copycat_vigscript(pkg, vig)
}
\arguments{
\item{pkg}{A search string}

\item{vig}{A search string}
}
\value{
A string
}
\description{
Create a new script with the code from the vignette.
}
