% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_vignettes.R
\name{copycat_vigsearch}
\alias{copycat_vigsearch}
\title{copycat_vigsearch}
\usage{
copycat_vigsearch(pkg)
}
\arguments{
\item{pkg}{A search string}
}
\value{
A list
}
\description{
copycat_vigsearch
}
