% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyCatCode.R
\docType{data}
\name{CopyCatCode}
\alias{CopyCatCode}
\title{Copy Cat Code:}
\format{
A data frame with 99 rows and 3 variables:
\describe{
  \item{package}{The R package name}
  \item{fct}{The function name}
  \item{code}{The code snippet}
  ...
}
}
\source{
\url{http://www.edgar-treischl.de/}
}
\usage{
CopyCatCode
}
\description{
A data set containing minimal examples of code snippets
that run with implemented data.
}
\keyword{datasets}
