\name{functions}
\alias{addMapability}
\alias{addBins}
\alias{balancedCenter}
\alias{binGC}
\alias{binParams}
\alias{breakPointInt}
\alias{combineBins}
\alias{doCorrection}
\alias{getChrLength}
\alias{getChrPloidy}
\alias{getNumReads}
\alias{getSegs}
\alias{initRdClass}
\alias{listify}
\alias{loessCorrect}
\alias{logScore}
\alias{methAdjust}
\alias{ploidyPercentages}
\alias{plotWindows}
\alias{readEntrypoints}
\alias{returnOutliers}
\alias{rpois.od}
\alias{stripOutliers}
\alias{binMap}
\alias{getAnnoDir}
\alias{getMedianReadCount}
\alias{makeCorrBins}
\alias{makeCorrBinsMedian}
\alias{makeDf}
\alias{makeDfLog}
\alias{makeDfLogPaired}
\alias{makeDfPaired}
\alias{makeMapDf}
\alias{plotGcLoess}
\alias{plotMapLoess}
\alias{replaceReadCounts}
\alias{addBins}
\alias{combinePercBins}
\alias{coverageFilter}
\alias{doSegPlot}
\alias{getWindowBins}
\alias{mergeLibraries}
\alias{plotSegsPaired}
\alias{plotSegsPairedLog}
\alias{removeGapCounts}
\alias{runPairedSampleAnalysis}
\alias{runSingleSampleAnalysis}
\alias{setParams}
\alias{subsetByReadLength}
\alias{writeBins}
\alias{writePairedBins}
\title{
  readDepth Functions
}
\description{
  miscellaneous internal functions used by the readDepth package
}
