% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSegments.R
\name{getSegs}
\alias{getSegs}
\title{Title}
\usage{
getSegs(
  rdo,
  gd2,
  params,
  entrypoints,
  onlyAlts,
  minWidth = 3,
  alpha = 0.01,
  rmGaps = TRUE,
  undoSD = 2
)
}
\arguments{
\item{undoSD}{}
}
\value{

}
\description{
Title
}
