% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_copy.R
\name{git_plot}
\alias{git_plot}
\title{git_plot}
\usage{
git_plot(file)
}
\arguments{
\item{file}{A character vector.}
}
\value{
A character vector.
}
\description{
git_plot
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit(x, split = ",")
}
