% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_copy.R
\name{git_search}
\alias{git_search}
\title{git_search}
\usage{
git_search(author, repository, branch = "master")
}
\arguments{
\item{author}{A character vector.}

\item{repository}{A character vector.}

\item{branch}{A character vector.}
}
\value{
A character vector.
}
\description{
git_search
}
