% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{loessCorrect}
\alias{loessCorrect}
\title{Title}
\usage{
loessCorrect(
  rdo,
  libNum,
  readLength,
  libName,
  outlierPercentage = 0.01,
  corrResolution = 0.001,
  type = "gc"
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Title
}
