% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotGcLoess}
\alias{plotGcLoess}
\title{Title}
\usage{
plotGcLoess(
  med,
  gc,
  reads,
  fitted,
  fixed,
  fixedfit,
  rdo,
  libNum,
  readLength,
  libName
)
}
\arguments{
\item{libName}{}
}
\value{

}
\description{
Title
}
