% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotWindows}
\alias{plotWindows}
\title{Title}
\usage{
plotWindows(
  windowSize,
  genomeSize,
  divGain,
  divLoss,
  fdr,
  numReads,
  oDisp,
  ploidyPerc,
  med
)
}
\arguments{
\item{med}{}
}
\value{

}
\description{
Title
}
