% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{runPairedSampleAnalysis}
\alias{runPairedSampleAnalysis}
\title{Title}
\usage{
runPairedSampleAnalysis(
  annotationDirectory,
  outputDirectory,
  normal,
  tumor,
  inputType,
  maxCores = 0,
  binSize = 0,
  gcWindowSize = 100,
  fdr = 0.01,
  perLibrary = TRUE,
  perReadLength = TRUE,
  readLength = 0,
  verbose = TRUE,
  outputSingleSample = FALSE,
  tumorSamtoolsFile = NULL,
  normalSamtoolsFile = NULL,
  samtoolsFileFormat = "10colPileup",
  dumpBins = FALSE,
  minWidth = 3,
  doGcCorrection = TRUE,
  rDataFile = NULL,
  minMapability = 0.6,
  purity = 1,
  maxGapOverlap = 0.75
)
}
\arguments{
\item{maxGapOverlap}{}
}
\value{

}
\description{
Title
}
