% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{runSingleSampleAnalysis}
\alias{runSingleSampleAnalysis}
\title{Title}
\usage{
runSingleSampleAnalysis(
  annotationDirectory,
  outputDirectory,
  inputFile,
  inputType,
  maxCores = 0,
  binSize = 0,
  gcWindowSize = 100,
  fdr = 0.01,
  perLibrary = TRUE,
  perReadLength = TRUE,
  readLength = 0,
  samtoolsFile = NULL,
  samtoolsFileFormat = "10colPileup",
  peakwiggle = 3,
  snpBinSize = 1e+06,
  verbose = TRUE,
  purity = 1
)
}
\arguments{
\item{purity}{}
}
\value{

}
\description{
Title
}
