% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{setParams}
\alias{setParams}
\title{Title}
\usage{
setParams(
  rdo,
  annotationDirectory,
  outputDirectory,
  inputFile,
  inputType,
  maxCores = 0,
  binSize = 0,
  gcWindowSize = 100,
  fdr = 0.01,
  perLibrary = TRUE,
  perReadLength = TRUE,
  readLength = 0,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
