% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{ask_gpt}
\alias{ask_gpt}
\title{ASK ChatGPT for some R help}
\usage{
ask_gpt(
  message,
  model = "text-davinci-003",
  maxtoken = 300,
  tempvalue = 0,
  return = "text"
)
}
\arguments{
\item{message}{The prompt for ChatGPT}

\item{model}{The ChatGPT model}

\item{maxtoken}{Maximal numbers and tokens}

\item{tempvalue}{The temperature}

\item{return}{Returns answer as text via the console (default) or raw as a character vector}
}
\value{
A character vector.
}
\description{
The function `ask_gpt` sends message to ChatGPT and shows the results via the console
}
