% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{cnNeutralDepthFromHetSites}
\alias{cnNeutralDepthFromHetSites}
\title{Title}
\usage{
cnNeutralDepthFromHetSites(
  rdo,
  samtoolsFile,
  snpBinSize,
  peakWiggle = 3,
  minimumDepth = 20,
  maximumDepth = 100,
  plot = FALSE,
  samtoolsFileFormat = "10colPileup",
  verbose = FALSE
)
}
\arguments{
\item{plot}{}

\item{verbose}{}
}
\value{



}
\description{
Title
}
\examples{
Title

}
