% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSegments.R
\name{cnSegments.paired}
\alias{cnSegments.paired}
\title{Title}
\usage{
cnSegments.paired(
  rdo.ref,
  rdo.test,
  onlyAlts = FALSE,
  minWidth = 3,
  alpha = 0.01,
  undoSD = 2,
  rmGaps = TRUE
)
}
\arguments{
\item{rmGaps}{}
}
\value{

}
\description{
Title
}
