% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSegments.R
\name{cnSegments}
\alias{cnSegments}
\title{Title}
\usage{
cnSegments(
  rdo,
  onlyAlts = FALSE,
  minWidth = 3,
  alpha = 0.01,
  undoSD = 2,
  rmGaps = TRUE
)
}
\arguments{
\item{rmGaps}{}
}
\value{

}
\description{
Title
}
