\name{copyCat-package}
\docType{package}
\title{
   Uses the depth of short reads in windows across the genome to 
   identify regions of genomic copy-number gain and loss
}
\description{
  Supports multi-core systems with the help of the foreach and doMC
  packages, degrades gracefully to single core systems.
}
\details{
  \tabular{ll}{
    Package: \tab copyCat\cr
    Type: \tab Package\cr
    Version: \tab 2.0\cr
    Date: \tab 2013-03-18\cr
    License: \tab Apache\cr
    LazyLoad: \tab yes\cr
  }
  

  ~~ An overview of how to use the package, including the most important ~~
  ~~ functions ~~
}
\author{
  Chris Miller
  
  Maintainer: Chris Miller <chrisamiller@gmail.com>
}
\references{
  TODO
}
\keyword{ package }
\examples{
#TODO
}
