% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_addin.R
\name{copycat_addin}
\alias{copycat_addin}
\title{Manage code snippets via the `copycat_addin()`.}
\usage{
copycat_addin(data = CopyCatCode)
}
\arguments{
\item{data}{A data frame}
}
\description{
The `copycat_addin()` creates a shiny gadget
to work with the Copycat or your data for Copycat. Pick an R package
and a function within the gadget. Press the insert
button and the code will be insert into the current document at the
location of the cursor. The `copycat_addin()` is inspired by parsnip addin,
which inserts model specifications.
}
