% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_code.R
\name{copycat_code}
\alias{copycat_code}
\title{copycat_code}
\usage{
copycat_code(x, data = CopyCatCode)
}
\arguments{
\item{x}{A search string for the function}

\item{data}{A data frame}
}
\value{
A text string that shows the code
}
\description{
`copycat_code()` uses a text string to search
for a code snippet. The function just returns that code.
It expects that you a data frame with three columns. A column
with the package name (package), a function (fun), and one with
the code (code) to copy. If data = NULL, the package search in the
the internal data `CopyCatCode`.
}
