% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_code.R
\name{copycat_code}
\alias{copycat_code}
\title{Inspect code from CopyCat.}
\usage{
copycat_code(x, data = CopyCatCode)
}
\arguments{
\item{x}{A search string for the function}

\item{data}{A data frame}
}
\value{
A text string that shows the code
}
\description{
The `copycat_code()` function searches for a
code snippet, but it returns only the code.
It expects a data frame with three columns. A column
with the package name (package), the function (fun) name, and
the code (code) to copy. As long as you do not provide a data frame, the function
search witing the `CopyCatCode` data.
}
