% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist.R
\name{copycat_gist}
\alias{copycat_gist}
\title{Explore Gist files}
\usage{
copycat_gist(filename)
}
\arguments{
\item{filename}{The file name (see copycat_gistfiles)}
}
\value{
Message if code was copied.
}
\description{
The function `copycat_gist()` copies GitHub Gists.
}
