% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{copycat_helpcode}
\alias{copycat_helpcode}
\title{Copy the help code examples}
\usage{
copycat_helpcode(pkg, fn)
}
\arguments{
\item{pkg}{A search string}

\item{fn}{A search string}
}
\value{
A string
}
\description{
The function copies
the code from the help file (examples) of an R package
}
