% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_help.R
\name{copycat_helpsearch}
\alias{copycat_helpsearch}
\title{copycat_helpsearch}
\usage{
copycat_helpsearch(pkg)
}
\arguments{
\item{pkg}{A search string}
}
\value{
A list
}
\description{
copycat_helpsearch
}
