% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_help.R
\name{copycat_helpsearch}
\alias{copycat_helpsearch}
\title{Examine the help files of a package.}
\usage{
copycat_helpsearch(pkg)
}
\arguments{
\item{pkg}{A search string}
}
\value{
A list
}
\description{
Examine the help files of a package.
}
