% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycat_code.R
\name{copycat_package}
\alias{copycat_package}
\title{copycat_package}
\usage{
copycat_package(x, data = CopyCatCode)
}
\arguments{
\item{x}{A search string}

\item{data}{Data}
}
\value{
A string
}
\description{
copycat_package
}
