% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{copycat_runhelp}
\alias{copycat_runhelp}
\title{Run code from the help files}
\usage{
copycat_runhelp(pkg, fn)
}
\arguments{
\item{pkg}{A search string}

\item{fn}{A search string}
}
\value{
A string
}
\description{
The function runs code from the help file (examples)
and sends it to the console
}
