\name{functions}
\alias{addMapability}
\alias{balancedCenter}
\alias{binGC}
\alias{binParams}
\alias{breakPointInt}
\alias{combineBins}
\alias{doCorrection}
\alias{getChrLength}
\alias{getChrPloidy}
\alias{getNumReads}
\alias{getSegs}
\alias{initRdClass}
\alias{listify}
\alias{loessCorrect}
\alias{logScore}
\alias{methAdjust}
\alias{ploidyPercentages}
\alias{plotWindows}
\alias{readEntrypoints}
\alias{returnOutliers}
\alias{rpois.od}
\alias{stripOutliers}
\alias{binMap}
\alias{getAnnoDir}
\alias{getMedianReadCount}
\alias{makeCorrBins}
\alias{makeCorrBinsMedian}
\alias{makeDf}
\alias{makeDfLog}
\alias{makeDfLogPaired}
\alias{makeDfPaired}
\alias{makeMapDf}
\alias{plotGcLoess}
\alias{plotMapLoess}
\alias{sort.data.frame}
\alias{replaceReadCounts}
\title{
  readDepth Functions
}
\description{
  miscellaneous internal functions used by the readDepth package
}
