% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{mapCorrect}
\alias{mapCorrect}
\title{Title}
\usage{
mapCorrect(
  rdo,
  outlierPercentage = 0.01,
  minMapability = 0.6,
  resolution = 0.01,
  skipCorrection = FALSE
)
}
\arguments{
\item{skipCorrection}{}
}
\value{

}
\description{
Title
}
