% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSegments.R
\name{removeCoverageArtifacts}
\alias{removeCoverageArtifacts}
\title{Title}
\usage{
removeCoverageArtifacts(segs, rdo)
}
\arguments{
\item{rdo}{}

\item{sp}{}
}
\value{



}
\description{
Title
}
\examples{
Title

}
