% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSegments.R
\name{removeGapSpanningSegments}
\alias{removeGapSpanningSegments}
\title{Title}
\usage{
removeGapSpanningSegments(segs, rdo, maxOverlap = 0.75, gapExpansion = 1)
}
\arguments{
\item{gapExpansion}{}
}
\value{

}
\description{
Title
}
